+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- `FocusTrap`:
+  - feat: Add className to wrapper element
+  - fix: Exclude `FocusTrap` elements from list of eligible focusables
 
 ## 0.10.125